backend/requirements.txt
fastapi
uvicorn
python-dotenv
dashscope
google-genai
Pillow
